unified styling

+++ src/components/GenericForm.js
@@ -1,18 +1,12 @@
 import React, { useState, useEffect } from "react";
-import {
-  View,
-  Text,
-  TextInput,
-  Button,
-  StyleSheet,
-  Alert,
-  Switch,
-} from "react-native";
+import { View, Text, TextInput, Button, Alert, Keyboard } from "react-native";
 import { useNavigation, useRoute } from "@react-navigation/native";
 import DateTimePicker from "@react-native-community/datetimepicker";
 import DropDownPicker from "react-native-dropdown-picker";
-import Ionicons from 'react-native-vector-icons/Ionicons';
+import Ionicons from "react-native-vector-icons/Ionicons";
+import { CheckBox } from "react-native-elements";
 import { addOrUpdateEntry, deleteEntry } from "../firebase/firestoreHelper";
+import { commonStyles } from "../styles/styles";
 
 const GenericForm = () => {
   const navigation = useNavigation();
@@ -25,11 +19,10 @@ const GenericForm = () => {
     duration: item?.duration || "",
     calories: item?.calories || "",
     date: item?.date ? new Date(item.date) : new Date(),
-    isSpecial: item?.isSpecial || false,
+    isSpecial: item?.isSpecial || undefined,
   });
   const [showDatePicker, setShowDatePicker] = useState(false);
 
-  // DropDownPicker state
   const [open, setOpen] = useState(false);
   const [value, setValue] = useState(formData.description);
   const [items, setItems] = useState([
@@ -42,8 +35,32 @@ const GenericForm = () => {
     { label: "Hiking", value: "Hiking" },
   ]);
 
-  const handleInputChange = (name, value) => {
+  const handleInputChange = async (name, value) => {
     setFormData((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleSpecialChange = (newValue) => {
+    if (!newValue) {
+      Alert.alert(
+        "Important",
+        "Are you sure you want to save these changes?",
+        [
+          {
+            text: "No",
+            onPress: () => console.log("No changes made."),
+            style: "cancel",
+          },
+          {
+            text: "Yes",
+            onPress: () =>
+              setFormData((prev) => ({ ...prev, isSpecial: newValue })),
+          },
+        ],
+        { cancelable: false }
+      );
+    } else {
+      setFormData((prev) => ({ ...prev, isSpecial: newValue }));
+    }
   };
 
   const handleSubmit = async () => {
@@ -55,32 +72,62 @@ const GenericForm = () => {
     const duration = parseInt(formData.duration);
     const calories = parseInt(formData.calories);
 
+    // Automatically determine if the entry should be special
+    const automaticIsSpecial =
+      collectionName === "diets" ? calories > 800 : duration > 60;
+
+    // If isSpecial has not been manually set by the user, use the automatic determination
+    const isSpecial =
+      formData.isSpecial !== undefined
+        ? formData.isSpecial
+        : automaticIsSpecial;
+
     const entryData = {
       description: formData.description,
       date: formData.date,
-      isSpecial: collectionName === "diets" ? calories > 800 : duration > 60,
-      ...(collectionName === "diets" ? { calories } : { duration })
+      isSpecial: isSpecial,
+      ...(collectionName === "diets" ? { calories } : { duration }),
     };
 
     const success = await addOrUpdateEntry(collectionName, entryData, item?.id);
     if (success) {
       Alert.alert("Success", "Entry updated successfully.");
-      navigation.goBack();
+      if (navigation.canGoBack()) {
+        navigation.goBack();
+      }
     } else {
       Alert.alert("Error", "Failed to update entry.");
     }
   };
 
-  const handleDelete = async () => {
-    if (item?.id) {
-      const success = await deleteEntry(collectionName, item.id);
-      if (success) {
-        Alert.alert("Success", "Entry deleted successfully.");
-        navigation.goBack();
-      } else {
-        Alert.alert("Error", "Failed to delete entry.");
-      }
-    }
+  const handleDelete = () => {
+    Alert.alert(
+      "Delete", // Title of the alert
+      "Are you sure you want to delete this item?", // Message of the alert
+      [
+        {
+          text: "No",
+          onPress: () => console.log("Deletion cancelled"),
+          style: "cancel",
+        },
+        {
+          text: "Yes",
+          onPress: async () => {
+            if (item?.id) {
+              const success = await deleteEntry(collectionName, item.id);
+              if (success) {
+                Alert.alert("Success", "Entry deleted successfully.");
+                if (navigation.canGoBack()) {
+                  navigation.goBack();
+                }
+              } else {
+                Alert.alert("Error", "Failed to delete entry.");
+              }
+            }
+          },
+        },
+      ]
+    );
   };
 
   useEffect(() => {
@@ -89,7 +136,7 @@ const GenericForm = () => {
 
   useEffect(() => {
     navigation.setOptions({
-      headerRight: () => (
+      headerRight: () =>
         isEditMode && (
           <Ionicons
             name="trash-bin"
@@ -98,19 +145,18 @@ const GenericForm = () => {
             style={{ marginRight: 15 }}
             onPress={handleDelete}
           />
-        )
-      ),
+        ),
     });
   }, [navigation, isEditMode]);
 
   return (
-    <View style={styles.container}>
-      <Text>{collectionName === "diets" ? "Description" : "Activity"}:</Text>
+    <View style={commonStyles.container}>
+      <Text>{collectionName === "diets" ? "Description " : "Activity "}*</Text>
       {collectionName === "diets" ? (
         <TextInput
           value={formData.description}
           onChangeText={(text) => handleInputChange("description", text)}
-          style={styles.input}
+          style={commonStyles.input}
         />
       ) : (
         <DropDownPicker
@@ -121,31 +167,41 @@ const GenericForm = () => {
           setValue={setValue}
           setItems={setItems}
           containerStyle={{ height: 40, marginBottom: 20 }}
-          style={{ backgroundColor: 'white' }}
-          dropDownContainerStyle={{ backgroundColor: 'white' }}
-          labelStyle={{ color: 'black' }}
+          style={{ backgroundColor: "white" }}
+          dropDownContainerStyle={{ backgroundColor: "white" }}
+          labelStyle={{ color: "black" }}
           placeholder="Select an activity"
           onChangeValue={(value) => handleInputChange("description", value)}
         />
       )}
-      <Text>{collectionName === "diets" ? "Calories" : "Duration"}:</Text>
+      <Text>{collectionName === "diets" ? "Calories " : "Duration "}*</Text>
       <TextInput
-        value={collectionName === "diets" ? String(formData.calories) : String(formData.duration)}
-        onChangeText={(text) => handleInputChange(collectionName === "diets" ? "calories" : "duration", text)}
+        value={
+          collectionName === "diets"
+            ? String(formData.calories)
+            : String(formData.duration)
+        }
+        onChangeText={(text) =>
+          handleInputChange(
+            collectionName === "diets" ? "calories" : "duration",
+            text
+          )
+        }
         keyboardType="numeric"
-        style={styles.input}
+        style={commonStyles.input}
       />
-      <Text>Date:</Text>
+      <Text>Date *</Text>
       <TextInput
         value={formData.date.toDateString()}
         onFocus={() => setShowDatePicker(true)}
-        style={styles.input}
+        style={commonStyles.input}
+        onBlur={() => Keyboard.dismiss()}
       />
       {showDatePicker && (
         <DateTimePicker
           value={formData.date}
           mode="date"
-          display="default"
+          display="inline"
           onChange={(event, selectedDate) => {
             setShowDatePicker(false);
             handleInputChange("date", selectedDate || formData.date);
@@ -153,46 +209,28 @@ const GenericForm = () => {
         />
       )}
       {isEditMode && (
-        <View style={styles.switchContainer}>
-          <Text>Special Entry:</Text>
-          <Switch
-            value={formData.isSpecial}
-            onValueChange={(value) => handleInputChange("isSpecial", value)}
+        <View>
+          <Text>
+            This item is marked as special. Select the checkbox if you would
+            like to approve it.
+          </Text>
+          <CheckBox
+            checked={formData.isSpecial}
+            onPress={() => handleSpecialChange(!formData.isSpecial)}
+            tintColors={{ true: "blue", false: "black" }}
           />
         </View>
       )}
-      <View style={styles.buttonContainer}>
-      <Button title="Cancel" onPress={() => navigation.goBack()} color="grey" />
+      <View style={commonStyles.buttonContainer}>
+        <Button
+          title="Cancel"
+          onPress={() => navigation.goBack()}
+          color="grey"
+        />
         <Button title="Save" onPress={handleSubmit} />
       </View>
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-  },
-  input: {
-    height: 40,
-    borderColor: "gray",
-    borderWidth: 1,
-    marginBottom: 20,
-    paddingHorizontal: 10,
-    width: "100%",
-  },
-  switchContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "space-between",
-    marginBottom: 20,
-  },
-  buttonContainer: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginTop: 20,
-  },
-});
 
 export default GenericForm;
